requirements.txt
python-dotenv
pandas
psycopg2